# server-go/doc/swagger/swagger.yaml
basePath: /v1
definitions:
  controller.AddCommentRequest:
    properties:
      text:
        type: string
    required:
    - text
    type: object
...
  controller.CommentWithLike:
    properties:
      comment:
        $ref: '#/definitions/db.Comment'
      isLikedMe:
        type: boolean
      numLikes:
        type: integer
    type: object
...
  controller.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  controller.GetCommentsOfPostResponse:
    properties:
      commentWithLikes:
        items:
          $ref: '#/definitions/controller.CommentWithLike'
        type: array
    type: object
  controller.GetPostsByUserResponse:
    properties:
      postWithLikes:
        items:
          $ref: '#/definitions/controller.PostWithLike'
        type: array
    type: object
...
  controller.LikeResponse:
    properties:
      isLikedMe:
        type: boolean
      numLikes:
        type: integer
    type: object
  controller.PostWithLike:
    properties:
      isLikedMe:
        type: boolean
      numLikes:
        type: integer
      post:
        $ref: '#/definitions/db.Post'
    type: object
  controller.SuccessResponse:
    properties:
      message:
        type: string
    type: object
...
  db.Comment:
    properties:
      created_at:
        type: string
      id:
        type: integer
      post_id:
        type: integer
      text:
        type: string
      user_id:
        type: integer
    type: object
  db.Post:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      image_url:
        type: string
      title:
        type: string
      user_id:
        type: integer
    type: object
host: localhost:8000
info:
  contact: {}
  description: server-go
  title: Instagram-lite-copy API
  version: "1.0"
paths:
...
  /posts/{postID}:
    delete:
      parameters:
      - description: Post ID
        in: path
        name: postID
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Remove a post by ID
      tags:
      - Posts
  /posts/{postID}/comments:
    get:
      consumes:
      - application/json
      description: Get comments of a post with pagination
      parameters:
      - description: Post ID
        in: path
        name: postID
        required: true
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.GetCommentsOfPostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get comments of a post
      tags:
      - Comments
  /posts/{postID}/comments/{commentID}:
    delete:
      parameters:
      - description: Comment ID
        in: path
        name: commentID
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Remove a comment by ID
      tags:
      - Comments
  /posts/{postID}/comments/{commentID}/like:
    put:
      consumes:
      - application/json
      parameters:
      - description: Comment ID
        in: path
        name: commentID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Number of likes and isLikedMe
          schema:
            $ref: '#/definitions/controller.LikeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Like or dislike a comment
      tags:
      - Comments
  /posts/{postID}/comments/create:
    post:
      consumes:
      - application/json
      description: Add a new comment to a post
      parameters:
      - description: Post ID
        in: path
        name: postID
        required: true
        type: integer
      - description: Comment text
        in: body
        name: text
        required: true
        schema:
          $ref: '#/definitions/controller.AddCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.Comment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Add a new comment
      tags:
      - Comments
  /posts/{postID}/like:
    put:
      parameters:
      - description: Post ID
        in: path
        name: postID
        required: true
        type: integer
      responses:
        "200":
          description: Number of likes and isLikedMe
          schema:
            $ref: '#/definitions/controller.LikeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Like or dislike a post
      tags:
      - Posts
  /posts/create:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Title of the post
        in: formData
        name: title
        type: string
      - description: Description of the post
        in: formData
        name: description
        type: string
      - description: Image of the post
        in: formData
        name: img
        required: true
        type: file
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.Post'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Add a new post
      tags:
      - Posts
  /posts/userID/{userID}:
    get:
      description: Get posts by user with pagination
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: List of posts with number likes and isLikedMe
          schema:
            $ref: '#/definitions/controller.GetPostsByUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get posts by user
      tags:
      - Posts
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"